+++ program/system/global.c
@@ -55,7 +55,7 @@ system_t system = {
 
 status_t SD_status;
 
-uint64_t boot_time = 0;
+uint32_t boot_time = 0;
 
 xSemaphoreHandle serial_tx_wait_sem = NULL;
 xQueueHandle serial_rx_queue = NULL;
